feat: fix auth issues
